Fix convention violations from code review (PIO-51)
- Add SenderIdentityResource for consistent serialisation layer
- Split SenderIdentityTest (611 lines) into SenderIdentityTest and
  SenderIdentitySecretSnapshotTest to stay under 500-line limit
- Eager-load senderIdentity on User model to avoid N+1 on every request
- Add user_id to SenderIdentityFactory default state
- Remove redundant identityType computed in SenderIdentityForm.vue
