+++ app/Http/Controllers/ConfigurationController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UpdateConfigurationRequest;
+use App\Http\Resources\SenderIdentityResource;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
@@ -16,14 +17,7 @@ class ConfigurationController extends Controller
 
         return Inertia::render('Settings/Index', [
             'storeMaskedRecipientEmail' => (bool) $request->user()->store_masked_recipient_email,
-            'senderIdentity' => $senderIdentity ? [
-                'type' => $senderIdentity->type,
-                'company_name' => $senderIdentity->company_name,
-                'domain' => $senderIdentity->domain,
-                'email' => $senderIdentity->email,
-                'verification_token' => $senderIdentity->verification_token,
-                'is_verified' => $senderIdentity->isVerified(),
-            ] : null,
+            'senderIdentity' => $senderIdentity ? new SenderIdentityResource($senderIdentity) : null,
             'planSupportsSenderIdentity' => $request->user()->planSupportsSenderIdentity(),
         ]);
     }
